fix: update gix to 0.66 for Rust compatibility
The old gix 0.38.0 had compilation errors with modern Rust versions
due to type annotation issues in gix-url. Updated to gix 0.66 which
is compatible with the latest stable Rust.

Changes:
- Update gix from 0.38.0 to 0.66
- Fix type cast: bytes.len() -> bytes.len() as u64 for loose_header
- Update Cargo.lock with all new dependencies
- Revert Docker image name back to original (lts-debian)

This resolves compilation errors like:
- "type annotations needed" in gix-url
- "mismatched types" expecting u64 instead of usize

+++ src/hasher.rs
@@ -46,7 +46,7 @@ pub fn git_hash(file_set: DashSet<PathBuf>, cwd: &str) -> Option<HashMap<String,
 
     if let Ok(bytes) = bytes_results {
       let mut hasher = hasher(gix::hash::Kind::Sha1);
-      hasher.update(&loose_header(gix::objs::Kind::Blob, bytes.len()));
+      hasher.update(&loose_header(gix::objs::Kind::Blob, bytes.len() as u64));
       hasher.update(&bytes);
 
       map.insert(key, Some(hex::encode(hasher.digest())));
@@ -75,7 +75,7 @@ pub fn git_hash_vec(files: Vec<PathBuf>, cwd: &str) -> Option<HashMap<String, Op
 
     if let Ok(bytes) = bytes_results {
       let mut hasher = hasher(gix::hash::Kind::Sha1);
-      hasher.update(&loose_header(gix::objs::Kind::Blob, bytes.len()));
+      hasher.update(&loose_header(gix::objs::Kind::Blob, bytes.len() as u64));
       hasher.update(&bytes);
 
       map.insert(key, Some(hex::encode(hasher.digest())));
